DesktopApp/requirements.txt
colorthief==0.2.1
darkdetect==0.8.0
numpy==2.1.3
pillow==11.0.0
PyQt-Fluent-Widgets==1.7.1
PyQt5==5.15.11
PyQt5-Frameless-Window==0.4.3
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
pywin32==308
scipy==1.14.1
